src/main/java/com/github/brunomndantas/jscrapper/support/parser/single/text/primitive/SinglePrimitiveByteTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.single.text.primitive;

import com.github.brunomndantas.jscrapper.support.parser.single.text.SingleTextParser;
import org.openqa.selenium.WebElement;

public class SinglePrimitiveByteTextParser extends SingleTextParser {

    public SinglePrimitiveByteTextParser() {
        super((byte)0);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Byte.parseByte(element.getText());
    }

}
